Use a lifetimed buffer internally
