+++ esp-hal/src/dma/buffers/mod.rs
@@ -10,6 +10,9 @@ use crate::soc::is_slice_in_dram;
 #[cfg(dma_can_access_psram)]
 use crate::soc::{is_slice_in_psram, is_valid_psram_address, is_valid_ram_address};
 
+pub(crate) mod scoped;
+pub(crate) use scoped::*;
+
 /// Error returned from Dma[Rx|Tx|RxTx]Buf operations.
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
@@ -495,11 +498,7 @@ pub struct BufView<T>(T);
 /// FIFO. See [DmaRxBuf] for receiving data.
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct DmaTxBuf {
-    descriptors: DescriptorSet<'static>,
-    buffer: &'static mut [u8],
-    burst: BurstConfig,
-}
+pub struct DmaTxBuf(ScopedDmaTxBuf<'static>);
 
 impl DmaTxBuf {
     /// Creates a new [DmaTxBuf] from some descriptors and a buffer.
@@ -514,7 +513,7 @@ impl DmaTxBuf {
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        Self::new_with_config(descriptors, buffer, BurstConfig::default())
+        ScopedDmaTxBuf::new_with_config(descriptors, buffer, BurstConfig::default()).map(Self)
     }
 
     /// Creates a new [DmaTxBuf] from some descriptors and a buffer.
@@ -530,80 +529,28 @@ impl DmaTxBuf {
         buffer: &'static mut [u8],
         config: impl Into<BurstConfig>,
     ) -> Result<Self, DmaBufError> {
-        let mut buf = Self {
-            descriptors: DescriptorSet::new(descriptors)?,
-            buffer,
-            burst: BurstConfig::default(),
-        };
-
-        let capacity = buf.capacity();
-        buf.configure(config, capacity)?;
-
-        Ok(buf)
-    }
-
-    fn configure(
-        &mut self,
-        burst: impl Into<BurstConfig>,
-        length: usize,
-    ) -> Result<(), DmaBufError> {
-        let burst = burst.into();
-        self.set_length_fallible(length, burst)?;
-
-        self.descriptors.link_with_buffer(
-            self.buffer,
-            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
-        )?;
-
-        self.burst = burst;
-        Ok(())
+        ScopedDmaTxBuf::new_with_config(descriptors, buffer, config).map(Self)
     }
 
     /// Configures the DMA to use burst transfers to access this buffer.
     pub fn set_burst_config(&mut self, burst: BurstConfig) -> Result<(), DmaBufError> {
-        let len = self.len();
-        self.configure(burst, len)
+        self.0.set_burst_config(burst)
     }
 
     /// Consume the buf, returning the descriptors and buffer.
     pub fn split(self) -> (&'static mut [DmaDescriptor], &'static mut [u8]) {
-        (self.descriptors.into_inner(), self.buffer)
+        self.0.split()
     }
 
     /// Returns the size of the underlying buffer
     pub fn capacity(&self) -> usize {
-        self.buffer.len()
+        self.0.capacity()
     }
 
     /// Return the number of bytes that would be transmitted by this buf.
     #[allow(clippy::len_without_is_empty)]
     pub fn len(&self) -> usize {
-        self.descriptors
-            .linked_iter()
-            .map(|d| d.len())
-            .sum::<usize>()
-    }
-
-    fn set_length_fallible(&mut self, len: usize, burst: BurstConfig) -> Result<(), DmaBufError> {
-        if len > self.capacity() {
-            return Err(DmaBufError::BufferTooSmall);
-        }
-        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::Out)?;
-
-        self.descriptors.set_tx_length(
-            len,
-            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
-        )?;
-
-        // This only needs to be done once (after every significant length change) as
-        // Self::prepare sets Preparation::auto_write_back to false.
-        for desc in self.descriptors.linked_iter_mut() {
-            // In non-circular mode, we only set `suc_eof` for the last descriptor to signal
-            // the end of the transfer.
-            desc.reset_for_tx(desc.next.is_null());
-        }
-
-        Ok(())
+        self.0.len()
     }
 
     /// Reset the descriptors to only transmit `len` amount of bytes from this
@@ -612,7 +559,7 @@ impl DmaTxBuf {
     /// The number of bytes in data must be less than or equal to the buffer
     /// size.
     pub fn set_length(&mut self, len: usize) {
-        unwrap!(self.set_length_fallible(len, self.burst))
+        self.0.set_length(len);
     }
 
     /// Fills the TX buffer with the bytes provided in `data` and reset the
@@ -621,18 +568,22 @@ impl DmaTxBuf {
     /// The number of bytes in data must be less than or equal to the buffer
     /// size.
     pub fn fill(&mut self, data: &[u8]) {
-        self.set_length(data.len());
-        self.as_mut_slice()[..data.len()].copy_from_slice(data);
+        self.0.fill(data);
     }
 
     /// Returns the buf as a mutable slice than can be written.
     pub fn as_mut_slice(&mut self) -> &mut [u8] {
-        self.buffer
+        self.0.as_mut_slice()
     }
 
     /// Returns the buf as a slice than can be read.
     pub fn as_slice(&self) -> &[u8] {
-        self.buffer
+        self.0.as_slice()
+    }
+
+    /// Consumes the buffer and returns the scoped version.
+    pub(crate) fn into_scoped(self) -> ScopedDmaTxBuf<'static> {
+        self.0
     }
 }
 
@@ -641,29 +592,7 @@ unsafe impl DmaTxBuffer for DmaTxBuf {
     type Final = DmaTxBuf;
 
     fn prepare(&mut self) -> Preparation {
-        cfg_if::cfg_if! {
-            if #[cfg(dma_can_access_psram)] {
-                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
-                if is_data_in_psram {
-                    unsafe {
-                        crate::soc::cache_writeback_addr(
-                            self.buffer.as_ptr() as u32,
-                            self.buffer.len() as u32,
-                        )
-                    };
-                }
-            }
-        }
-
-        Preparation {
-            start: self.descriptors.head(),
-            direction: TransferDirection::Out,
-            #[cfg(dma_can_access_psram)]
-            accesses_psram: is_data_in_psram,
-            burst_transfer: self.burst,
-            check_owner: None,
-            auto_write_back: false,
-        }
+        self.0.prepare()
     }
 
     fn into_view(self) -> BufView<DmaTxBuf> {
@@ -682,11 +611,7 @@ unsafe impl DmaTxBuffer for DmaTxBuf {
 /// See [DmaTxBuf] for transmitting data.
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct DmaRxBuf {
-    descriptors: DescriptorSet<'static>,
-    buffer: &'static mut [u8],
-    burst: BurstConfig,
-}
+pub struct DmaRxBuf(ScopedDmaRxBuf<'static>);
 
 impl DmaRxBuf {
     /// Creates a new [DmaRxBuf] from some descriptors and a buffer.
@@ -700,7 +625,7 @@ impl DmaRxBuf {
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        Self::new_with_config(descriptors, buffer, BurstConfig::default())
+        ScopedDmaRxBuf::new_with_config(descriptors, buffer, BurstConfig::default()).map(Self)
     }
 
     /// Creates a new [DmaRxBuf] from some descriptors and a buffer.
@@ -716,70 +641,29 @@ impl DmaRxBuf {
         buffer: &'static mut [u8],
         config: impl Into<BurstConfig>,
     ) -> Result<Self, DmaBufError> {
-        let mut buf = Self {
-            descriptors: DescriptorSet::new(descriptors)?,
-            buffer,
-            burst: BurstConfig::default(),
-        };
-
-        buf.configure(config, buf.capacity())?;
-
-        Ok(buf)
-    }
-
-    fn configure(
-        &mut self,
-        burst: impl Into<BurstConfig>,
-        length: usize,
-    ) -> Result<(), DmaBufError> {
-        let burst = burst.into();
-        self.set_length_fallible(length, burst)?;
-
-        self.descriptors.link_with_buffer(
-            self.buffer,
-            burst.max_chunk_size_for(self.buffer, TransferDirection::In),
-        )?;
-
-        self.burst = burst;
-        Ok(())
+        ScopedDmaRxBuf::new_with_config(descriptors, buffer, config).map(Self)
     }
 
     /// Configures the DMA to use burst transfers to access this buffer.
     pub fn set_burst_config(&mut self, burst: BurstConfig) -> Result<(), DmaBufError> {
-        let len = self.len();
-        self.configure(burst, len)
+        self.0.set_burst_config(burst)
     }
 
     /// Consume the buf, returning the descriptors and buffer.
     pub fn split(self) -> (&'static mut [DmaDescriptor], &'static mut [u8]) {
-        (self.descriptors.into_inner(), self.buffer)
+        self.0.split()
     }
 
     /// Returns the size of the underlying buffer
     pub fn capacity(&self) -> usize {
-        self.buffer.len()
+        self.0.capacity()
     }
 
     /// Returns the maximum number of bytes that this buf has been configured to
     /// receive.
     #[allow(clippy::len_without_is_empty)]
     pub fn len(&self) -> usize {
-        self.descriptors
-            .linked_iter()
-            .map(|d| d.size())
-            .sum::<usize>()
-    }
-
-    fn set_length_fallible(&mut self, len: usize, burst: BurstConfig) -> Result<(), DmaBufError> {
-        if len > self.capacity() {
-            return Err(DmaBufError::BufferTooSmall);
-        }
-        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::In)?;
-
-        self.descriptors.set_rx_length(
-            len,
-            burst.max_chunk_size_for(&self.buffer[..len], TransferDirection::In),
-        )
+        self.0.len()
     }
 
     /// Reset the descriptors to only receive `len` amount of bytes into this
@@ -788,25 +672,22 @@ impl DmaRxBuf {
     /// The number of bytes in data must be less than or equal to the buffer
     /// size.
     pub fn set_length(&mut self, len: usize) {
-        unwrap!(self.set_length_fallible(len, self.burst));
+        self.0.set_length(len)
     }
 
     /// Returns the entire underlying buffer as a slice than can be read.
     pub fn as_slice(&self) -> &[u8] {
-        self.buffer
+        self.0.as_slice()
     }
 
     /// Returns the entire underlying buffer as a slice than can be written.
     pub fn as_mut_slice(&mut self) -> &mut [u8] {
-        self.buffer
+        self.0.as_mut_slice()
     }
 
     /// Return the number of bytes that was received by this buf.
     pub fn number_of_received_bytes(&self) -> usize {
-        self.descriptors
-            .linked_iter()
-            .map(|d| d.len())
-            .sum::<usize>()
+        self.0.number_of_received_bytes()
     }
 
     /// Reads the received data into the provided `buf`.
@@ -815,31 +696,18 @@ impl DmaRxBuf {
     /// first `buf.len()` bytes of received data is written into `buf`.
     ///
     /// Returns the number of bytes in written to `buf`.
-    pub fn read_received_data(&self, mut buf: &mut [u8]) -> usize {
-        // Note that due to an ESP32 quirk, the last received descriptor may not get
-        // updated.
-        let capacity = buf.len();
-        for chunk in self.received_data() {
-            if buf.is_empty() {
-                break;
-            }
-            let to_fill;
-            (to_fill, buf) = buf.split_at_mut(chunk.len());
-            to_fill.copy_from_slice(chunk);
-        }
-
-        capacity - buf.len()
+    pub fn read_received_data(&self, buf: &mut [u8]) -> usize {
+        self.0.read_received_data(buf)
     }
 
     /// Returns the received data as an iterator of slices.
     pub fn received_data(&self) -> impl Iterator<Item = &[u8]> {
-        self.descriptors.linked_iter().map(|desc| {
-            // SAFETY: We set up the descriptor to point to a subslice of the buffer, and
-            // here we are only recreating that slice with a perhaps shorter length.
-            // We are also not accessing `self.buffer` while this slice is alive, so we
-            // are not violating any aliasing rules.
-            unsafe { core::slice::from_raw_parts(desc.buffer.cast_const(), desc.len()) }
-        })
+        self.0.received_data()
+    }
+
+    /// Consumes the buffer and returns the scoped version.
+    pub(crate) fn into_scoped(self) -> ScopedDmaRxBuf<'static> {
+        self.0
     }
 }
 
@@ -848,34 +716,7 @@ unsafe impl DmaRxBuffer for DmaRxBuf {
     type Final = DmaRxBuf;
 
     fn prepare(&mut self) -> Preparation {
-        for desc in self.descriptors.linked_iter_mut() {
-            desc.reset_for_rx();
-        }
-
-        cfg_if::cfg_if! {
-            if #[cfg(dma_can_access_psram)] {
-                // Optimization: avoid locking for PSRAM range.
-                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
-                if is_data_in_psram {
-                    unsafe {
-                        crate::soc::cache_invalidate_addr(
-                            self.buffer.as_ptr() as u32,
-                            self.buffer.len() as u32,
-                        )
-                    };
-                }
-            }
-        }
-
-        Preparation {
-            start: self.descriptors.head(),
-            direction: TransferDirection::In,
-            #[cfg(dma_can_access_psram)]
-            accesses_psram: is_data_in_psram,
-            burst_transfer: self.burst,
-            check_owner: None,
-            auto_write_back: true,
-        }
+        self.0.prepare()
     }
 
     fn into_view(self) -> BufView<DmaRxBuf> {
